by ravi

+++ visog-job-portal-ui/WebContent/resources/js/jobRole.js
@@ -69,12 +69,14 @@
 
 						// HTTP DELETE- delete jobRole by Id
 						$scope.deleteJobRole = function(jobRole) {
+							if(confirm("Are you sure to Delete ?") == true){
 							$http(
 									{
 										method : 'DELETE',
 										url : 'http://localhost:8080/visog-job-portal-api/master/jobRole/'
 												+ jobRole.id
 									}).then(_success, _error);
+							}
 						};
 
 						// In case of edit, populate form fields and assign form.id
